refactor: trail lib folder name
