+++ src/libs/trail/types.ts
